splitting filetree tab moved mod info dialog classes to a sub filter
