+++ src/modinfodialog.ui
@@ -20,7 +20,7 @@
       <enum>QTabWidget::Rounded</enum>
      </property>
      <property name="currentIndex">
-      <number>7</number>
+      <number>8</number>
      </property>
      <property name="movable">
       <bool>true</bool>
@@ -1088,7 +1088,7 @@ p, li { white-space: pre-wrap; }
        </item>
       </layout>
      </widget>
-     <widget class="QWidget" name="tabFiles">
+     <widget class="QWidget" name="tabFiles1">
       <attribute name="title">
        <string>Filetree</string>
       </attribute>
@@ -1096,7 +1096,7 @@ p, li { white-space: pre-wrap; }
        <item>
         <layout class="QHBoxLayout" name="horizontalLayout_13">
          <item>
-          <widget class="QPushButton" name="openInExplorerButton">
+          <widget class="QPushButton" name="openInExplorerButton1">
            <property name="sizePolicy">
             <sizepolicy hsizetype="Preferred" vsizetype="Fixed">
              <horstretch>0</horstretch>
@@ -1124,7 +1124,7 @@ p, li { white-space: pre-wrap; }
         </layout>
        </item>
        <item>
-        <widget class="QTreeView" name="fileTree">
+        <widget class="QTreeView" name="fileTree1">
          <property name="contextMenuPolicy">
           <enum>Qt::CustomContextMenu</enum>
          </property>
